backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "poetry.lock"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.1.1"
jax = "^0.4.26"
optax = "^0.2.2"
matplotlib = "^3.8.4"
tqdm = "^4.66.2"
django = "^5.0.4"
click = "^8.1.7"
djangorestframework = "^3.15.1"
pre-commit = "^3.7.0"
mypy-extensions = "^1.0.0"
packaging = "^24.0"
pathspec = "^0.12.1"
pip = "^24.0"
platformdirs = "^4.2.0"
tk = "^0.1.0"
tzdata = "^2024.1"
wheel = "^0.43.0"
setuptools = "^69.5.1"
requests = "^2.31.0"
django-cors-headers = "^4.3.1"
django-environment = "^0.3.8"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2 = "^2.9.9"
psycopg = "^3.1.18"
django-storages = "^1.14.2"
boto3 = "^1.34.92"


[build-system]
requires = ["poetry-core"]
package-mode = false
build-backend = "poetry.core.masonry.api"
